INTRODUCED models

+++ src/main/java/org/alienlabs/hatchetharry/view/component/gui/ReorderCardInBattlefieldBehavior.java
@@ -1,5 +1,14 @@
 package org.alienlabs.hatchetharry.view.component.gui;
 
+import java.io.IOException;
+import java.math.BigInteger;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.UUID;
+
+import javax.servlet.http.HttpServletRequest;
+
 import org.alienlabs.hatchetharry.HatchetHarrySession;
 import org.alienlabs.hatchetharry.model.MagicCard;
 import org.alienlabs.hatchetharry.model.channel.ReorderCardCometChannel;
@@ -20,14 +29,6 @@ import org.apache.wicket.util.template.TextTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Required;
-
-import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
-import java.math.BigInteger;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.UUID;
 
 public class ReorderCardInBattlefieldBehavior extends AbstractDefaultAjaxBehavior
 {
@@ -100,7 +101,7 @@ public class ReorderCardInBattlefieldBehavior extends AbstractDefaultAjaxBehavio
 				mc.setBattlefieldOrder(Integer.valueOf(index));
 			}
 			card.setBattlefieldOrder(newIndex);
-			allCardsInBattlefieldForPlayer.add(newIndex, card);
+			allCardsInBattlefieldForPlayer.add(newIndex.intValue(), card);
 		}
 
 		Collections.sort(allCardsInBattlefieldForPlayer);
@@ -144,8 +145,9 @@ public class ReorderCardInBattlefieldBehavior extends AbstractDefaultAjaxBehavio
 		}
 		catch (final IOException e)
 		{
-			ReorderCardInBattlefieldBehavior.LOGGER.error(
-					"unable to close template in ReorderCardInBattlefieldBehavior#renderHead()!", e);
+			ReorderCardInBattlefieldBehavior.LOGGER
+					.error("unable to close template in ReorderCardInBattlefieldBehavior#renderHead()!",
+							e);
 		}
 	}
